test(vllm): isolate ACK checks from NPU suite
Keep CPU-only socket coverage out of the accelerator-wide NPU process to avoid its teardown crash.
