+++ tests/utils/test_bucketed_weight_transfer.py
@@ -125,36 +125,6 @@ def test_sender_accepts_strided_tensor(monkeypatch):
     assert socket.poll_calls == [(sender.ack_timeout_ms, bucketed_weight_transfer.zmq.POLLIN)]
 
 
-def test_sender_ack_supports_recv_only_socket():
-    from verl.workers.rollout.vllm_rollout import bucketed_weight_transfer
-
-    class _RecvOnlySocket:
-        def recv(self):
-            return b""
-
-    sender = bucketed_weight_transfer.BucketedWeightSender("ipc:///tmp/test-bwt-unused.sock")
-    sender.socket = _RecvOnlySocket()
-
-    sender._receive_ack("test")
-
-
-def test_sender_ack_timeout_raises():
-    from verl.workers.rollout.vllm_rollout import bucketed_weight_transfer
-
-    class _TimeoutSocket:
-        def poll(self, _timeout, _flags):
-            return 0
-
-        def recv(self):
-            raise AssertionError("recv must not be called after a timeout")
-
-    sender = bucketed_weight_transfer.BucketedWeightSender("ipc:///tmp/test-bwt-unused.sock", ack_timeout_ms=1)
-    sender.socket = _TimeoutSocket()
-
-    with pytest.raises(RuntimeError, match="timed out waiting 1ms"):
-        sender._receive_ack("test")
-
-
 # ---------------------------------------------------------------------------
 # Process entry points (must be module-level for pickling with spawn)
 # ---------------------------------------------------------------------------
